requirements.txt
django==4.2
segno
numpy 
gunicorn
python-barcode
pillow
aspose-barcode-for-python-via-net 
